[refacotring] ConverInput name변경
